extract shared query logic to db/queries module, import from route and server function
Moves getMeetingByBodyAndDateQuery and FiscalDecisionDetail type from StorageService to new db/queries.ts module. Updates StorageService to import and re-export only MeetingDetail type. Updates route component and server function to import directly from db/queries. Removes duplicate FiscalDecision type alias in route component.
